+++ linux-3.10.79/fs/yaffs2/yaffs_getblockinfo.h
@@ -1,7 +1,8 @@
 /*
  * YAFFS: Yet another Flash File System . A NAND-flash specific file system.
  *
- * Copyright (C) 2002-2018 Aleph One Ltd.
+ * Copyright (C) 2002-2011 Aleph One Ltd.
+ *   for Toby Churchill Ltd and Brightstar Engineering
  *
  * Created by Charles Manning <charles@aleph1.co.uk>
  *
@@ -22,8 +23,7 @@
 static inline struct yaffs_block_info *yaffs_get_block_info(struct yaffs_dev
 							      *dev, int blk)
 {
-	if (blk < (int)dev->internal_start_block ||
-	    blk > (int)dev->internal_end_block) {
+	if (blk < dev->internal_start_block || blk > dev->internal_end_block) {
 		yaffs_trace(YAFFS_TRACE_ERROR,
 			"**>> yaffs: get_block_info block %d is not valid",
 			blk);
